feat: clean up create transaction command and add outfile option
